Issue #1438, refactored pgp password
