+++ extension/js/common/core/pgp.ts
@@ -10,7 +10,6 @@ import { MsgBlock, MsgBlockType, Mime } from './mime.js';
 import { AttMeta } from './att.js';
 import { mnemonic } from './mnemonic.js';
 import { requireOpenpgp } from '../platform/require.js';
-import { secureRandomBytes, base64encode } from '../platform/util.js';
 import { FcAttLinkData } from './att.js';
 import { Buf } from './buf.js';
 import { Xss } from '../platform/xss.js';
@@ -171,18 +170,6 @@ export enum DecryptErrTypes {
   needPassphrase = 'need_passphrase',
   format = 'format',
   other = 'other',
-}
-
-export interface PasswordStrengthResult {
-  word: {
-    match: string;
-    word: string;
-    bar: number;
-    color: string;
-    pass: boolean;
-  };
-  seconds: number;
-  time: string;
 }
 
 export class FormatError extends Error {
@@ -501,28 +488,6 @@ export class Pgp {
       }
       throw new Error('No valid signature found in key');
     }
-  };
-
-  public static password = {
-    estimateStrength: (zxcvbnResultGuesses: number, type: 'passphrase' | 'pwd' = 'passphrase'): PasswordStrengthResult => {
-      const timeToCrack = zxcvbnResultGuesses / Pgp.CRACK_GUESSES_PER_SECOND;
-      for (const word of type === 'pwd' ? Pgp.CRACK_TIME_WORDS_PWD : Pgp.CRACK_TIME_WORDS_PASS_PHRASE) {
-        const readableTime = Pgp.internal.readableCrackTime(timeToCrack);
-        if (readableTime.includes(word.match)) { // looks for a word match from readable_crack_time, defaults on "weak"
-          return { word, seconds: Math.round(timeToCrack), time: readableTime };
-        }
-      }
-      throw Error('(thrown) estimate_strength: got to end without any result');
-    },
-    weakWords: () => [
-      'crypt', 'up', 'cryptup', 'flow', 'flowcrypt', 'encryption', 'pgp', 'email', 'set', 'backup', 'passphrase', 'best', 'pass', 'phrases', 'are', 'long', 'and', 'have', 'several',
-      'words', 'in', 'them', 'Best pass phrases are long', 'have several words', 'in them', 'bestpassphrasesarelong', 'haveseveralwords', 'inthem',
-      'Loss of this pass phrase', 'cannot be recovered', 'Note it down', 'on a paper', 'lossofthispassphrase', 'cannotberecovered', 'noteitdown', 'onapaper',
-      'setpassword', 'set password', 'set pass word', 'setpassphrase', 'set pass phrase', 'set passphrase'
-    ],
-    random: () => { // eg TDW6-DU5M-TANI-LJXY
-      return base64encode(openpgp.util.Uint8Array_to_str(secureRandomBytes(128))).toUpperCase().replace(/[^A-Z0-9]|0|O|1/g, '').replace(/(.{4})/g, '$1-').substr(0, 19);
-    },
   };
 
   public static internal = {
